feat: implement blog and portfolio pages with client components and enhanced data fetching
- Added BlogPageClient and PortfolioPageClient components for improved structure and reusability.
- Updated blog and portfolio pages to utilize async data fetching for blogs, tags, and projects.
- Enhanced image handling in the Next.js configuration for better performance.
- Implemented pagination and filtering features in the portfolio and blog components.
- Cleaned up and organized code for better maintainability.
